fix resend qos1 test

+++ src/Subscription.cpp
@@ -4,7 +4,7 @@
 
 #include "Subscription.h"
 
-Subscription::Subscription(char *topic, long int lastMsgIdConsumed, unsigned short qos) : topic(topic),
+Subscription::Subscription(char *topic, int lastMsgIdConsumed, unsigned short qos) : topic(topic),
                                                                                                last_msg_id_consumed(
                                                                                                        lastMsgIdConsumed),
                                                                                                qos(qos) {}
@@ -13,7 +13,7 @@ char *Subscription::getTopic() const {
     return topic;
 }
 
-long int Subscription::getLastMsgIdConsumed() const {
+int Subscription::getLastMsgIdConsumed() const {
     return last_msg_id_consumed;
 }
 
@@ -21,6 +21,6 @@ unsigned short Subscription::getQos() const {
     return qos;
 }
 
-void Subscription::setLastMsgIdConsumed(long int lastMsgIdConsumed) {
+void Subscription::setLastMsgIdConsumed(int lastMsgIdConsumed) {
     last_msg_id_consumed = lastMsgIdConsumed;
 }
