+++ src/packets/PublishAckPacket.cpp
@@ -4,9 +4,9 @@
 
 #include "PublishAckPacket.h"
 
-PublishAckPacket::PublishAckPacket(RawPacket *rawPacket, long int packetId) : RawPacket(rawPacket),
+PublishAckPacket::PublishAckPacket(RawPacket *rawPacket, int packetId) : RawPacket(rawPacket),
                                                                                    packetId(packetId) {}
 
-long int PublishAckPacket::getPacketId() const {
+int PublishAckPacket::getPacketId() const {
     return packetId;
 }
