+++ src/topic/Topic.h
@@ -10,10 +10,10 @@ class Topic {
 
 
 public:
-    long int _last_msg_id_published;
-    long int _subscribed_users_count;
+    int _last_msg_id_published;
+    int _subscribed_users_count;
     char *value;
-    Topic(long int lastMsgIdPublished, long int subscribedUserCount, char *topic);
+    Topic(int lastMsgIdPublished, int subscribedUserCount, char *topic);
     Topic(char* topic);
 
     long getLastMsgIdPublished() const;
